refactor login.js and introduce userSession.js
